Update AS5600.cpp
Critical bug fix. The I2CADDR and I2CUPDT registers are single byte registers on AS5600L, without specifying this the default behaviour is to write two bytes. The address will then be mistakenly written as 0x00 which is a reserved I2C address. If this is followed by a burn settings, the wrong address may permanently be written and the sensor may be 'bricked'.